Raise 403 on missing headers + Don't raise KeyError when "type" key is missing from request (#3)
* REMOVED superfluous pass

* UPDATED __init__.py: Don't error out when request does contain 'type'

If 'type' is not in the request, `.get()` will be `None` (which != `InteractionType.PING`) which avoids the KeyError that would occur otherwise.

* UPDATED __init__.py: Don't call `verify_key()` if headers are not present

Taking advantage of the shortcircuiting of `or`, whenever either of the `X-Signature-*` headers is absent, `verify_key` will not be called and the 401 status will be returned. This avoids the `AttributeError` that would occur when line 24 tries to `.encode()` `None`.